requirements.txt
sklearn
cycler==0.10.0
Django==2.1.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
matplotlib==2.2.3
mongoengine==0.15.3
mysqlclient==1.3.13
numpy==1.15.1
pkg-resources==0.0.0
PyJWT==1.6.4
pymongo==3.7.1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.5
scikit-learn==0.19.2
scipy==1.1.0
six==1.11.0
